reorganize mocks in rmc
