+++ rmc/Cargo.toml
@@ -20,5 +20,6 @@ futures-timer = "3.0"
 log = "0.4"
 
 [dev-dependencies]
+aleph-bft-mock = { path = "../mock" }
 rand = "0.8"
 tokio = { version = "1.6.1", features = ["macros", "rt", "rt-multi-thread"] }
